+++ src/react/src/components/ProductionLocation/ContributionsDrawer/ContributionCard/ContributionCard.jsx
@@ -6,6 +6,7 @@ import {
     instanceOf,
     bool,
     number,
+    node,
 } from 'prop-types';
 import { Link } from 'react-router-dom';
 import Typography from '@material-ui/core/Typography';
@@ -35,6 +36,7 @@ const ContributionCard = ({
         data-testid={dataTestId}
     >
         <Typography
+            component="div"
             className={
                 promoted
                     ? `${classes.contributionValue} ${classes.contributionValuePromoted}`
@@ -106,7 +108,7 @@ const ContributionCard = ({
 
 ContributionCard.propTypes = {
     classes: object.isRequired,
-    value: string.isRequired,
+    value: oneOfType([string, node]).isRequired,
     sourceName: string,
     date: oneOfType([string, instanceOf(Date)]),
     promoted: bool,