+++ src/react/src/components/ProductionLocation/ProductionLocationDetailsMap/utils.js
@@ -4,9 +4,7 @@ import partition from 'lodash/partition';
 import uniqBy from 'lodash/uniqBy';
 
 import { STATUS_CLAIMED, STATUS_CROWDSOURCED } from '../DataPoint/constants';
-import { FIELD_TYPE } from './constants';
-
-export { FIELD_TYPE };
+import { FIELD_CONFIG } from '../constants';
 
 export const getContributorStatus = (contributorName, isFromClaim) => {
     if (!contributorName) return null;
@@ -26,7 +24,7 @@ export const getContributorStatus = (contributorName, isFromClaim) => {
  */
 export const getFieldContributorInfo = (singleFacilityData, fieldType) => {
     switch (fieldType) {
-        case FIELD_TYPE.ADDRESS: {
+        case FIELD_CONFIG.address.key: {
             const address =
                 get(singleFacilityData, 'properties.address', '') || '';
             const addressFields = get(
@@ -96,10 +94,19 @@ export const getFieldContributorInfo = (singleFacilityData, fieldType) => {
                 })),
             };
 
-            return { contributorName, userId, date, status, drawerData };
+            return {
+                key: 'address',
+                label: FIELD_CONFIG.address.label,
+                tooltipText: FIELD_CONFIG.address.tooltipText,
+                contributorName,
+                userId,
+                date,
+                status,
+                drawerData,
+            };
         }
 
-        case FIELD_TYPE.COORDINATES: {
+        case FIELD_CONFIG.coordinates.key: {
             const coordinates = get(
                 singleFacilityData,
                 'geometry.coordinates',
@@ -194,11 +201,23 @@ export const getFieldContributorInfo = (singleFacilityData, fieldType) => {
 
             const drawerData = { promotedContribution, contributions };
 
-            return { contributorName, userId, date, status, drawerData };
+            return {
+                key: 'coordinates',
+                label: FIELD_CONFIG.coordinates.label,
+                tooltipText: FIELD_CONFIG.coordinates.tooltipText,
+                contributorName,
+                userId,
+                date,
+                status,
+                drawerData,
+            };
         }
 
         default:
             return {
+                key: null,
+                label: null,
+                tooltipText: null,
                 contributorName: '',
                 userId: null,
                 date: '',